Add PiStorm utilities / drivers HDF
A small hard drive image that includes all the utilities and drivers for
PiStorm. Along with a shell script to rebuild it.

It is mounted by default using piscsi6 when piscsi is enabled.

This branch also fixes the GUI crash that can happen when an interrupt
fires during a ROM switch.