Incredibly questionable virtual AHI device support
To test this, install AHI (probably preferably 4.18) and copy pi-ahi.audio to DEVS:AHI and PI-AHI to DEVS:AudioModes.
The IRQ trigger magically doesn't work, so this all runs of vertical blank... meaning that it stops working pretty much the instant you try to use it in an RTG mode.
Not very useful at the moment, but I need to take a break from working on this to do a bunch of work.